Echobot version 1.0
Simple while True and respond function mechanism. You enter something and if your sentence is in the dictionary the echobot will return one of the (random chosen) pre-made responses. If the echobot does not know (not added to the dictionary) what you are saying it will return ""I can hear you! you said: [message]".
If you ask "How old are you?" it will calculate how many days it has been since the echobot was created